extract common parts from scripts
