+++ scripts/_common.ts
@@ -1,0 +1,67 @@
+import fs from "fs"
+
+import chalk from "chalk"
+import * as A from "fp-ts/lib/Array"
+import { log } from "fp-ts/lib/Console"
+import * as IO from "fp-ts/lib/IO"
+import * as T from "fp-ts/lib/Task"
+import * as TE from "fp-ts/lib/TaskEither"
+import { Endomorphism } from "fp-ts/lib/function"
+import { pipe } from "fp-ts/lib/pipeable"
+import * as glob from "glob"
+
+export const readFile = TE.taskify<fs.PathLike, string, NodeJS.ErrnoException, string>(
+  fs.readFile
+)
+
+export const writeFile = TE.taskify<fs.PathLike, string, NodeJS.ErrnoException, void>(
+  fs.writeFile
+)
+
+const exit = (code: 0 | 1): IO.IO<void> => () => process.exit(code)
+
+export function onLeft(e: NodeJS.ErrnoException): T.Task<void> {
+  return T.fromIO(
+    pipe(
+      log(e),
+      IO.chain(() => exit(1))
+    )
+  )
+}
+
+export function onRight(msg: string) {
+  return (): T.Task<void> => T.fromIO(log(chalk.bold.green(msg)))
+}
+
+function modifyFile(
+  f: Endomorphism<string>
+): (path: string) => TE.TaskEither<NodeJS.ErrnoException, void> {
+  return (path) =>
+    pipe(
+      readFile(path, "utf8"),
+      TE.map((original) => ({ original, updated: f(original) })),
+      TE.chain(({ original, updated }) =>
+        original === updated ? TE.of(undefined) : writeFile(path, updated)
+      )
+    )
+}
+
+function modifyFiles(
+  f: Endomorphism<string>
+): (paths: Array<string>) => TE.TaskEither<NodeJS.ErrnoException, void> {
+  return (paths) =>
+    pipe(
+      A.array.traverse(TE.taskEither)(paths, modifyFile(f)),
+      TE.map(() => undefined)
+    )
+}
+
+export function modifyGlob(
+  f: Endomorphism<string>
+): (pattern: string) => TE.TaskEither<NodeJS.ErrnoException, void> {
+  return (pattern) => pipe(glob.sync(pattern), TE.right, TE.chain(modifyFiles(f)))
+}
+
+export function runMain(t: T.Task<void>): Promise<void> {
+  return t().catch((e) => console.log(chalk.bold.red(`Unexpected error: ${e}`)))
+}
